Corrected preinstall script to post hook
The preinstall script was trying to run webpack, but webpack hasn't
installed yet with that hook. I've updated it to be a post hook.

I also made some improvements to lessen redundancy in the package.json.
The post hook was only echoing finished which isn't as elegant as simply
exiting cleanly without an error code.